backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
mammoth==1.6.0
aiofiles==23.2.1